moved visualizer into classes folder
