+++ client/src/classes/AudioVisualizer.ts
